refactor: migrate to portaphone app identification
