+++ android/app/google-services.json
@@ -1,43 +1,45 @@
 {
   "project_info": {
-    "project_number": "973049735976",
-    "project_id": "webtrit-app",
-    "storage_bucket": "webtrit-app.appspot.com"
+    "project_number": "524910534863",
+    "project_id": "webtrit-69559",
+    "storage_bucket": "webtrit-69559.firebasestorage.app"
   },
   "client": [
     {
       "client_info": {
-        "mobilesdk_app_id": "1:973049735976:android:3b8774082803ad8f547e1d",
+        "mobilesdk_app_id": "1:524910534863:android:c9fed097ac3546684ad7ba",
         "android_client_info": {
           "package_name": "com.webtrit.app"
         }
       },
-      "oauth_client": [
-        {
-          "client_id": "973049735976-p4dg2dikj1goumh7i00l0pf6d628s66q.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
+      "oauth_client": [],
       "api_key": [
         {
-          "current_key": "AIzaSyAq1cn3u6oeWVHTdAAhwdsy78tRpb4ab2Y"
+          "current_key": "AIzaSyBiHZTx8_6-kdgVXwKoIxgfPAXL7ug6tVQ"
         }
       ],
       "services": {
         "appinvite_service": {
-          "other_platform_oauth_client": [
-            {
-              "client_id": "973049735976-p4dg2dikj1goumh7i00l0pf6d628s66q.apps.googleusercontent.com",
-              "client_type": 3
-            },
-            {
-              "client_id": "973049735976-t0te56rlsu1putg89pt2b8c4qq8f80c1.apps.googleusercontent.com",
-              "client_type": 2,
-              "ios_info": {
-                "bundle_id": "com.webtrit.app"
-              }
-            }
-          ]
+          "other_platform_oauth_client": []
+        }
+      }
+    },
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:524910534863:android:8035f3b2f5b319c14ad7ba",
+        "android_client_info": {
+          "package_name": "com.webtrit.phone"
+        }
+      },
+      "oauth_client": [],
+      "api_key": [
+        {
+          "current_key": "AIzaSyBiHZTx8_6-kdgVXwKoIxgfPAXL7ug6tVQ"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": []
         }
       }
     }
