+++ android/app/src/main/kotlin/com/webtrit/phone/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.webtrit.app
+package com.webtrit.phone
 
 import io.flutter.embedding.android.FlutterActivity
