+++ dart_define.json
@@ -5,13 +5,13 @@
   "_WEBTRIT_APP_PREDEFINED_CORE_URLS": "",
   "__WEBTRIT_APP_PREDEFINED_CORE_URLS_DESCRIPTION": "Semicolon-separated list of Core URLs to pre-populate the core URL autocomplete dropdown on the login screen. Example: 'https://core1.example.com;https://core2.example.com'. Suggestions are always shown; user-typed entries appear above presets and case-insensitive duplicates are merged.",
   "_WEBTRIT_APP_CORE_VERSION_CONSTRAINT": "",
-  "WEBTRIT_APP_LINK_DOMAIN": "app.webtrit.com",
+  "WEBTRIT_APP_LINK_DOMAIN": "app.webtrit.phone",
   "__WEBTRIT_APP_LINK_DOMAIN_DESCRIPTION": "If provided, enables Android App Links integration via intent-filter. Must match assetlinks.json configuration on the target domain. This functionality is configured via Android flavors — if the value is non-empty, the 'deeplinks' flavor is selected automatically.",
   "WEBTRIT_APP_NAME": "WebTrit",
   "_WEBTRIT_APP_WEB_BUNDLE_ID": "",
   "__WEBTRIT_APP_WEB_BUNDLE_ID_DESCRIPTION": "Web only. Overrides the bundle_id sent in the session (login) and autoprovision requests on web, where there is no platform bundle identifier (packageInfo.packageName resolves to the pubspec name and the backend rejects it with 'unconfigured_bundle_id'). Disabled by default (leading underscore). To enable for a web build, remove the leading underscore and set this to a bundle_id that is registered on the server for the 'web' app type. Ignored on Android/iOS. Left disabled, web falls back to packageInfo.packageName.",
   "_WEBTRIT_APP_ABOUT_URL": "",
-  "WEBTRIT_APP_SALES_EMAIL": "sales@webtrit.com",
+  "WEBTRIT_APP_SALES_EMAIL": "sales@portaone.com",
   "_WEBTRIT_APP_FCM_VAPID_KEY": "",
   "__WEBTRIT_APP_FCM_VAPID_KEY_DESCRIPTION": "Optional public VAPID key used for Web Push notifications with Firebase Cloud Messaging (FCM). Required only for web platform to authenticate push subscription with FCM. Not used on mobile platforms (Android/iOS).",
   "WEBTRIT_APP_REMOTE_LOGZIO_LOG_LEVEL": "INFO",
