+++ ios/Runner/GoogleService-Info.plist
@@ -2,22 +2,18 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>CLIENT_ID</key>
-	<string>973049735976-t0te56rlsu1putg89pt2b8c4qq8f80c1.apps.googleusercontent.com</string>
-	<key>REVERSED_CLIENT_ID</key>
-	<string>com.googleusercontent.apps.973049735976-t0te56rlsu1putg89pt2b8c4qq8f80c1</string>
 	<key>API_KEY</key>
-	<string>AIzaSyDDRXt-svULnPAy7K_sMgC0tGB3Tj_PNFc</string>
+	<string>AIzaSyADReAZ-JT7QMiTtfhR-oVGfv42lqQwFEM</string>
 	<key>GCM_SENDER_ID</key>
-	<string>973049735976</string>
+	<string>524910534863</string>
 	<key>PLIST_VERSION</key>
 	<string>1</string>
 	<key>BUNDLE_ID</key>
-	<string>com.webtrit.app</string>
+	<string>com.webtrit.phone</string>
 	<key>PROJECT_ID</key>
-	<string>webtrit-app</string>
+	<string>webtrit-69559</string>
 	<key>STORAGE_BUCKET</key>
-	<string>webtrit-app.appspot.com</string>
+	<string>webtrit-69559.firebasestorage.app</string>
 	<key>IS_ADS_ENABLED</key>
 	<false></false>
 	<key>IS_ANALYTICS_ENABLED</key>
@@ -29,6 +25,6 @@
 	<key>IS_SIGNIN_ENABLED</key>
 	<true></true>
 	<key>GOOGLE_APP_ID</key>
-	<string>1:973049735976:ios:c8fdad3d4a4bc81a547e1d</string>
+	<string>1:524910534863:ios:ecdcf7348f2d1d914ad7ba</string>
 </dict>
 </plist>
